Add Docker configuration and update npm scripts
Added Docker and Docker Compose configuration files: Dockerfile and docker-compose.yml. This allows developers to set up the project in a reproducible environment. Made changes to package.json, updating the start script to allow for debugging directly in Docker. A bootstrap dependency was also added.  The .gitignore file has been updated to ignore the .pnpm-store, associated with the newly added pnpm package manager.