Split start script into start and debug
The start script in package.json now only runs the application. A new script, "debug", was added to package.json, which runs the application with debug options. This change allows developers to start the app without debugging if they prefer to do so.

In App.test.tsx, the test assertion has been commented out temporarily. This is likely due to an ongoing development, please remember to uncomment once the relevant feature is implemented.