Add Docker configuration and update npm scripts (#9)
* Add Docker configuration and update npm scripts

Added Docker and Docker Compose configuration files: Dockerfile and docker-compose.yml. This allows developers to set up the project in a reproducible environment. Made changes to package.json, updating the start script to allow for debugging directly in Docker. A bootstrap dependency was also added.  The .gitignore file has been updated to ignore the .pnpm-store, associated with the newly added pnpm package manager.

* Split start script into start and debug

The start script in package.json now only runs the application. A new script, "debug", was added to package.json, which runs the application with debug options. This change allows developers to start the app without debugging if they prefer to do so.

In App.test.tsx, the test assertion has been commented out temporarily. This is likely due to an ongoing development, please remember to uncomment once the relevant feature is implemented.

* Switch to debug mode in Docker compose

Changed the command in docker-compose.yml file for app service from "pnpm run start" to "pnpm run debug". This change allows us to have better debugging capabilities while running the app in a Docker container.