Remove old database model; Add initial PostgreSQL migration
The existing ephemeral database model snapshot and associated files have been removed as the project is being transitioned towards the use of a persistent PostgreSQL database. This commit includes the addition of the initial database migration for PostgreSQL, including the modelling of all primary domain entities for the application. Besides, the docker development configuration, example environment file, and nginx configuration have been added for aiding development and deployment.